Persist task across reinstalls
